Change the edit filter
